Enable WebGPU testing on Linux (#6591)
Enable testing the WebGPU backend on Linux using Chrome Dev (105) instead of Chrome Canary (even though base: 'ChromeCanary' is listed, it launches Chrome Dev on Linux). Fix incorrect passing of flags to Chrome. Flags enabled in chrome://flags need to appear between --flag-switches-begin and --flag-switches-end.

Note that not all tests are passing on Linux. This could be due to Chrome Dev being behind Chrome Canary, or due to a lack of testing on Linux before this PR.