Set willReadFrequently to true for GPU correctness test
If willReadFrequently is set to true, 2D canvas will always be backed by
CPU. Otherwise, it will be backed by GPU in first call, and backed by
CPU in the rest calls.
For CPU backends, including CPU and WASM, we set this to true for
perfomance. So to compare the results of GPU with CPU, we need to ensure
it's always backed by CPU.

BUG: #6664