De-flake fromPixels test failure
`HTMLVideoElement.requestVideoFrameCallback()` provides a realiable way
to guarantee the readiness of a video frame, whereas the browser may not
update the ready state until a few frames later. In such cases, false
alarms will be raised by the validation in `fromPixels()`.

For where rVFC is unavailable, fallback to the old path based on the
`loadeddata` event. This at least seems better than waiting on nothing.

Fixes #6577