Fix kernels_to_ops generating in the wrong location (#6823)
Avoid using `chdir = 'tfjs-converter'` in kernels_to_ops.bzl. Using `chdir` follows the symlink to `tfjs-converter`'s files in the execroot directory (`/usr/local/google/home/msoulanille/.cache/bazel/_bazel_msoulanille/f671279d7efd00bb0451adc485135cc0/execroot/tfjs/tfjs-converter`). Then, when node is started, it doesn't know that it followed a symlink, so when it tries to write to `../bazel-out/.....`, it's no longer looking in the execroot directory and can not find the correct file.

Fix this by starting node in the execroot directory instead of in the `tfjs-converter` symlink's directory.