Run verdaccio tests concurrently with bazel tests (#6999)
Previously, nightly Verdaccio tests ran exclusively with Bazel tests because it was as convenient way to make the rest of the build steps wait for Verdaccio to finish. This PR properly sets up the waitFor field of the other steps to depend on Verdaccio tests when they're present.

To do this, the PR adds a new waitedForByPackages flag to the cloudbuild config, which makes all non-bazel packages waitFor any step it's added to.