Make tfjs-tfdf work with bundlers (#7054)
Set the ESM entrypoint to the flat ESM bundle. Set the cjs entrypoint to the node (cjs) bundle.

Prevent node-specific 'require's in the emscripten bundle from being hoisted to the top of the file. Now, it checks if it's running in node before requiring node-specific things like 'fs' or 'path'.