Adds function overloads to minimize.
This will allow the caller to get correct type checking on the return values.

Note: This does not validate that the minimize function correctly returns the right type.  For example, if you changed `if (returnCost)` to `if (!returnCost)` you would not get a compiler error, so care must still be taken in ensuring this function does the right thing internally.

This change results in a number of tests failing to compile because they reuse variables (`let` instead of `const`).  I don't mind going through and fixing them all, but I wanted to get feedback on whether this is a desirable change first before I spend the time doing that.