Make converter hashtable test not overly specific
tfjs-converter's saved model conversion test for hashtable depends on the exact names of the hashtable weights it saves, but these names are generated and may change between versions of TF. This was observed when syncing into google3, which is using tf 2.12. This PR prevents the test from checking the name of the hashtable weights.