[webgpu] Prepare for binary op restructuring
The current code isn't great in that the vec4 shaders have diverged from
the scalar ones more than necessary. The situation will be much improved
by moving the NaN checking snippet in each op to a centralized place.
Here is the common preparation work, so that following refactoring can
be done on a per-op basis.