Remove async reshape from model analysis
Some backends will run part of the model on the CPU, and this depends on several
factors. When the subgraph which is an input to reshape is run on the CPU, then
the shape tensor's value is available synchronously, and the model can run with
`predict()`.

To avoid breaking this behavior, do not try to determine whether a reshape is
synchronous or asynchronous in model analysis at this point in time. It may be
possible later if more info about the current backend is available. For now,
though, rely on model execution's promise detection to throw an error if reshape
returns a promise during model.predict.