Never pass a NoOp value to an op
The 'NoOp' op appears in the execution graph (although it shouldn't really), and
its output value can sometimes be passed to downstream nodes. Temporarily fix
this by filtering it out of the parameters in `getParamValue`.

TODO(mattSoulanille): Refactor the graph so NoOp and other control dependencies
are only used for sorting and don't appear in the execution graph.