Support loading large model weights (#7610)
* webgpu: Fix a bug in softmax (#7607)

* Avoid allocating a large arraybuffer when loading weights (#7598)

The loadWeights function loads weights in 4MB chunks and then concatenates them into a single large ArrayBuffer. That ArrayBuffer is used for splitting the weights data back up into tensors. Allocating large ArrayBuffers (3.5GB) can be unstable on Chrome, so this PR avoids this allocation, instead slicing the weights out of the chunks manually.

The implementation wraps the array of weights (stored as ArrayBuffer[]) in a new CompositeArrayBuffer class. This class implements slice by copying the desired range out of the buffer(s) that it overlaps with.

* Support using a list of ArrayBuffers as model weight data

* Avoid 'Array.flat()'

* Simplify some of the tests

* Do not export 'CompositeArrayBuffer' from tfjs-core

* Update doc for weightData

* Fix tfjs-node

* Remove unused import

---------

Co-authored-by: Jiajia Qin <jiajia.qin@intel.com>