Fix golden integration tests not awaiting results (#7676)
Golden e2e model tests were not waiting for the asynchronous `tensor.data()` function to finish when checking results, making it hard to tell which test is failing. This PR fixes that.

After this fix, golden model tests are consistently timing out on BrowserStack Mac
Chrome. They work locally when tested on an M1 mac. Disable them on BrowserStack
for now.