webgpu: Replace timestamp-query-in-passes with timestamp-query
Timestamp-query is a core feature, which should be generally supported
on all platforms, including macOS. Note that Chrome switch
'--enable-dawn-features=allow_unsafe_apis' is still needed now as the
timestamp has the accuracy of nanosecond, which is too accurate to be
safe. Later changes in Chrome may lift this limitation.