webgpu: Tighten the condition to go the plain argminmax
In PR #6778, I ever added the plain argminmax shader to improve the perf
of ArgMax[1, 1025, 2049, 19]. However, I accidently added the output
size condition, which was not verified. This PR removes that condition
and only goes to plain argminmax when the reduce length is very small(<32).
With this change, ArgMax[1,513,513,151] becomes 7.19ms from 13.24ms in
DeepLabV3-ade20k on Intel ADL.